feat: 태그 선택 컴포넌트 분리
