+++ src/components/writePost/index.js
@@ -1,5 +1,6 @@
 export { default as RegisterPostButton } from "./RegisterPostButton";
 export { default as TagSelector } from "./TagSelector";
+export { default as TagSelectors } from "./TagSelectors";
 export { default as ImageUploader } from "./ImageUploader";
 export { default as PostTextarea } from "./PostTextarea";
 export { default as TopicModalContent } from "./bottomModal/TopicModalContent";
